+++ Tests/MusicCraftCoreTests/MusicCraftCoreTests.swift
@@ -3,6 +3,6 @@ import XCTest
 
 final class MusicCraftCoreTests: XCTestCase {
     func testVersionIsSet() {
-        XCTAssertEqual(musicCraftCoreVersion, "0.0.11")
+        XCTAssertEqual(musicCraftCoreVersion, "0.0.12")
     }
 }